+++ src/collective/documentgenerator/viewlets/batchactions.py
@@ -12,4 +12,3 @@ class DGBatchActionsViewlet(BatchActionsViewlet):
     def available(self):
         """Global availability of the viewlet."""
         return self.view.__name__ == 'dg-templates-listing'
-